pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "timetable-to-image"
version = "1.0.4"
description = "The Python library provides functionality for generating a timetable image from an object representation"
readme = "README.md"
authors = [{ name = "potapovns", email = "ns.potapov73@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["timetable", "image"]
dependencies = [
    "Pillow"
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = ["bumpver", "pytest"]

[project.urls]
Homepage = "https://github.com/potapovns/timetable_to_image"

[tool.bumpver]
current_version = "1.0.4"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"',
]
"src/TimetableToImage/__init__.py" = ["{version}"]

